Add initial simple integration with rich for rendering ValidationError's as a rich Tree
